Moved Page generation to separate file
Broke PageGeneration out into its own file and updated links. Bumped version to Alpha 3
