+++ cbzplatterlib/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
 __author__ = 'Mike Killian'
-__version__ = '0.0.2'
+__version__ = '0.0.3'
